requirements.txt
langcorn
pymongo
nltk
pytest
python-dotenv
gpt4all
pymongo
appdirs
black
click
dnspython
email-validator
fastapi
flake8
h11
idna
mccabe
motor
mypy-extensions
pathspec
pycodestyle
pydantic
pyflakes
regex
starlette
toml
typed-ast
typing-extensions
uvicorn